feature(AuthComponent): improved & renamed to AppAuth
