--- src/_REFACTOR_/context/Auth/AuthEvents.tsx
@@ -1,13 +1,0 @@
-export const AUTH_LOGGED_IN_EVENT = 'AUTH_LOGGED_IN_EVENT';
-export const AUTH_LOGGED_OUT_EVENT = 'AUTH_LOGGED_OUT_EVENT';
-
-interface AuthLoggedInEvent {
-  type: typeof AUTH_LOGGED_IN_EVENT;
-}
-
-interface AuthLoggedOutEvent {
-  type: typeof AUTH_LOGGED_OUT_EVENT;
-}
-
-export type AuthEventNames = typeof AUTH_LOGGED_OUT_EVENT | typeof AUTH_LOGGED_OUT_EVENT;
-export type AuthEventTypes = AuthLoggedInEvent | AuthLoggedOutEvent;
