--- src/_REFACTOR_/context/Auth/AuthState.ts
@@ -1,11 +1,0 @@
-import User from '../type/User';
-
-export interface AuthState {
-  subject: string;
-  user: User | null;
-}
-
-export const authDefaultState: AuthState = {
-  subject: '',
-  user: null
-};
