--- src/_REFACTOR_/context/type/Address.ts
@@ -1,5 +1,0 @@
-export interface Address {
-  id?: string;
-  country?: string;
-  city?: string;
-}
