--- src/_REFACTOR_/context/type/Profile.ts
@@ -1,6 +1,0 @@
-export interface Profile {
-  bio?: string;
-  location?: string;
-  avatarHash?: string;
-  profileLink?: string;
-}
